New release 0.1.4: 1. Move to_* function to to.py. 2. Add alternative period specifiers. E.g. instead of `s` now one can use `sec`. 3. Little refactoring for tests.
