+++ ptimedelta/const.py
@@ -1,0 +1,1 @@
+MS_IN_SEC = 1000
