+++ pyproject.toml
@@ -1,8 +1,8 @@
 [tool.poetry]
 name = "ptimedelta"
-version = "0.1.3"
+version = "0.1.4"
 description = "Pretty Timedelta. Convert your timedelta objects to pretty strings and vice versa."
-authors = ["Anton Goy <g.a.s.s@mail.ru>"]
+authors = ["Anton Goy <antongoy@internet.ru>"]
 license = "MIT"
 readme = "README.md"
 homepage = "https://github.com/antongoy/ptimedelta"
